[video_player] Simplify Pigeon API on Android (#7909)
The Pigeon API for `video_player_android` dates back to the earliest versions of Pigeon, when methods had to take only a single class parameter, and return a class. This updates the API surface to modern Pigeon style, eliminating wrapper classes (except for `CreateMessage`, which is complex enough that the class is useful).